requirements.txt
flask>=2.0
requests>=2.28
pyyaml>=6.0
jinja2>=3.0
pandas>=1.5
streamlit>=1.28
openai>=1.0
tiktoken>=0.5
python-dotenv>=1.0
